+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+### Added
+
+- Start chats from any directory using the `C-u` prefix for `pimacs-chat`,
+  which opens a transient for selecting a session name and root
+  directory.
+
 ## v0.1.0 - 2026-07-19
 
 ### Added